+++ lang/fr.go
@@ -6,4 +6,7 @@ var FR_STRINGS = LangString{
 	"output-desc":      "Répertoire de sortie",
 	"output-dir-empty": "Le répertoire de sortie est vide",
 	"silent":           "Désactiver l'impression du journal sur stdout",
+	"convert-usage":    "Convertir un scan en epub",
+	"no-chapter":       "Aucun chapitre spécifié",
+	"converting":       "Conversion de %d chapitres\n",
 }