+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 * Move main command in `commands/base.go`
+* Minor text changes
 
 ## [1.2.0] - 2024-05-07